requirements.txt
pandas
plotly
streamlit
scikit-learn
requests
python-dotenv